Rework telegram RX and TX mechanism
Previously, PDUs were not available for data telegram because it was
unclear what buffer to store them in and how to manage lifetime of this
buffer.

Rework the API of telegram RX/TX such that the PDU is accessed directly
from the PHY buffer.  Lifetime of this access is limited by only making
the telegram available inside a closure controlled by the PHY API.  Any
handling code must be moved inside of such a closure.

This design will allow zero-copy telegram send&receive which will be
very useful for low-memory devices.

One caveat of the new API is that the FDL master can no longer ensure
that it's `last_telegram_time` is always correctly updated.  It needs to
be manually updates using `mark_rx()` and `mark_tx()` now and these
calls may be easy to forget, leading to nasty timeout phenomena.  It may
be best to move the `last_telegram_time` into the PHY where we always
know for sure when the last transmission actually happened...