phy: simulator: Enforce synchronization pause
If the current token-wielding master wants to transmit, it needs to
adhere to the synchronization pause of 33 bit times.  Track the
token-owner in the simulator bus and ensure this participant always
waits for the required synchronization pause time before transmitting.
Other bus participants only need to wait for the 11 bit times minimum
Tsdr.