dp: Warn when too much peripheral storage is provided
As there are only 125 addresses and we occupy one, it never makes sense
to provide storage for more than 124 peripherals.  Warn when this is
done so users can fix their code.

We can think about making this a panic!() in the future, to really drive
home the point.  I am not sure whether this is a good idea or not.