fdl: Move watchdog time into FDL parameters
The watchdog timeout should be configured globally for the entire FDL
master instead of per peripheral.  Thus, make it an FDL parameter,
as that is where other PROFIBUS stacks keep it as well.

Additionally, store the pre-computed factors instead of the timeout
duration so we do not recompute these factors unnecessarily.