fdl: master: Implement proper slot timeout
The slot timeout must be the time between end of transmission and the
first byte received.  Implement it as such by leveraging the bus
activity tracking.

The current implementation comes with one caveat:  When an incomplete
telegram is received, the FDL master conservatively lets Tsl pass once
more before starting the next transmission.  It is unclear what behavior
is the correct one in this case.  This is also documented in a comment.