fdl: Drop Option<> around events type
This `Option<>` proves to be unergonomic.  Instead, require the `Events`
type to be `Default` and return its default value when no events exist.
This makes user code much more ergonomic.  If a particular application
does want an `Option`, it can do this by setting its `Events` type to an
`Option<T>`.

Generally, this slightly changes the meaning of the `Event` type: It
represents a collection of all `Events` that may have occurred.  To make
this more explicit, also rename the type from `Event` to `Events`.