phy: rp2040: Fix wrong cursor placement on TX
When first attempting transmission, the cursor was incorrectly placed.
`uart.write_raw()` returns a _remaining_ slice so the length of this
slice is _not_ the cursor location from the beginning, but rather the
cursor position from the end of the buffer.

Fix this by subtracting the length of the returned slice from the buffer
length to get the cursor position.  This was already handled correctly
in the `is_transmitting()` codepaths...