fdl: master: Fix a time race related to bus activity tracking
So far, the code relied on check_for_bus_activity() to detect incoming
data and mark bus activity accordingly.  However, a receive_telegram()
call later in the poll code may also receive new data which was so far
not accounted for in the bus activity timestamp.  This lead to violation
of the synchronization pause.

Fix this by explicitly updating the bus activity timestamp whenever a
telegram is received.

Alongside this, reset the `pending_bytes` marker on telegram receival
which allows us to get rid of the post-poll call to
`phy.get_pending_received_bytes()`.  This fixes another (less likely)
time race where bytes arrive between telegram receival and end of poll.