fdl: master: Ensure is_transmitting() is called during every poll cycle
The current ProfibusPhy trait API is a bit unintuitive.  PHY
implementations do a lot of actual work during the `is_transmitting()`
call - this is where data is usually shoveled into the hardware.

This becomes problematic when the FDL layer does not call
`is_transmitting()` during each and every poll cycle. This was
unintentionally broken in the `check_for_ongoing_transmision()`
implementation where `is_transmitting()` was only called after the
transmission time had already passed.

Fix this for now by moving the `is_transmitting()` call to the front of
the conditional.

For the future, a better API for the ProfibusPhy trait is needed. The
is_transmitting() call shouldn't be abused for performing non-obvious
work.  Instead, we want an explicit "work"/hardware-poll call which is
decoupled from this.