phy: Implement partial RX dropping
Previously, whenever a telegram was deserialized, the entire RX buffer
would be dropped.  This meant that only one telegram could ever be
received at a time.

This is problematic on systems with large poll latencies (e.g. Linux).
There, we need to robustly handle receiving multiple telegrams at once.
To make this work, only drop the deserialized telegram from the RX
buffer and keep the rest.

For now, this is only implemented for the Linux PHY.  The RP2040 PHY
will still perform full drops for the time being - on a real-time system
we shouldn't ever run into partial drops anyway.