examples: rp-pico: Implement a custom panic handler
To gain access to panic messages, implement a custom panic handler which
prints the panic message to a USB serial device.

The USB serial device is recreated in the panic handler which means it
needs to be reopened on the host.  Using a terminal emulator like `tio`
works best for this as it auto-reconnects to a tty device.