fdl: master: Check if token was received after forwarding it
After forwarding the token, we must check that the next master has
properly received it.  We can do this by listening for it to start
communication on the bus.

If we do not see communication, it means the token was not received.
For now, reacquire the token ourselves - in the future we must retry
forwarding the token here.