fdl: tests: Add test for a master dropping while in the ring
This test scenario simulates a bus master getting unplugged while it is
part of the token ring.  The other active stations must behave correctly
anyway.

In the future, the timing of this testcase should be tightened to ensure
the other active stations manage to resolve the situation in the
specifified time limit.