gsd-parser: Fix not parsing all keywords case-insensitive
Keywords need to be prefixed with `^` for pest to parse them
case-insensitively.  This is required by the GSD standard and we do see
some files out in the wild that deviate from the default spelling.

We already parse keywords case-insensitive on the rust side, this commit
fixes the pest grammar to follow suit.