gsd-parser: Correctly parse signed numbers where appropriate
The handling of different number types isn't really cleanly implemented
right now.  This lead to the oversight of not even allowing signed
integers in ExtUserPrmData where they should be.

As a first solution, parse all ExtUserPrmData values as signed integers
in the hopes that this is better than what we currently do...