fdl: master: Introduce RingState to track ring entrance progress
When entering the token ring, the FdlMaster must listen to two full
token rotations to learn about all active stations on the bus.

Introduce a new RingState which tracks progress in this step and
ensures the FdlMaster does not report readiness until it has completed
the listening phase.

This implementation does not yet track the list of active stations
(LAS).