Add a type alias for bus addresses
To make it more explicit in code where addresses are stored, define a
`profirust::Address` type alias that is used across the codebase to
reference bus addresses.

Additionally, define a method to check correctness of addresses.  This
can be used internally to verify that no incorrect addresses are passed
to profirust functions.