fdl: Drop event pass-through from FdlApplication design
In preparation for multi-application support, we unfortunately need to
drop the design where the application event is returned by the FDL
poll() method.  This cannot work with multiple applications present as
the event type will be different for each application.

Instead, FDL applications shall store their event internally and provide
a method to the user to retrieve the last event.  Implement this for all
our current applications using a `take_last_event()` design where the
method returns the last event only once.