dp: active: Factor out state data for UseToken into separate struct
Move the `token_time` into a struct that's meant to hold all state
relevant to the `UseToken` state.  This will make it more easy to add
additional state data in the future, as the entire struct can then be
passed around.

Passing around the data of the `UseToken` state is required to
transition into the `AwaitDataResponse` state and back.