fdl: active: Implement multi application polling
Add support for running an `FdlActiveStation` with multiple applications
on-top.  Cooperative round-robin scheduling is implemented to arbitrate bus time
between the applications.  Each application gets to complete its cycle
before the next application is scheduled.  The motivation for this is to
keep application telegrams temporally close together.

As a part of this, `FdlApplication` objects are now dispatched
dynamically throughout the entire `FdlActiveStation` implementation.
For single application usecases, the optimized should hopefully see
through the dyn and optimize regardless.  For multi-application
usecases, dynamic dispatch is required.