fdl: active: Fix incorrect bus activity tracking
During the rewrite of the active station, this essential check got
missed:  No matter what state we are in, unless we are currently
transmitting, we need to check the bus for new incoming data.  If
something arrived, we shall mark the current time, even if we did not
yet receive a full telegram.

Without calling `check_for_bus_activity()` we would only mark activity
on receival of a full telegram which would miss unluckily timed
responses that only get captured in half on slot timeout.

Add back the check_for_bus_activity() call as previously present.

Related to this, it may be sensible to _only_ perform actual PHY
receival at this one point early in the poll routine and then later only
refer to already received data.  To be evaluated...