phy: Add method for processing all pending telegrams at once
When running profirust in non-realtime contexts, there may be situations
where multiple telegrams are received at once.  This can happen when we
do not poll fast enough to catch each telegram in a separate poll cycle.

Previously, profirust would then start massively lagging behind by
delaying processing of an already received telegram until the next poll
cycle.

To mitigate this, add a new `receive_all_telegrams()` method to the PHY
interface which can be used to process all pending telegrams at once.
The caller is kept informed when it is processing a telegram that is not
the most recent one.